requirements.txt
iris-ued        >= 5.3.0, <6
scikit-ued      >= 2.1.3, <3
crystals        >= 1.3.0, <2
numpy           >= 1.20, <2
matplotlib      >= 3, <4
tqdm
scipy           >= 1.5, <2
scikit-image    >= 0.18, <1
termcolor       > 1, <2
